fix(social): lift search FAB to screen level so the offline indicator can't push it up
The Friends search FAB lived inside the Friend Requests TabView scene. The
scene is a flex sibling of `<OfflineIndicator style={{marginBottom:
bottomTabBarHeight}} />`, so the indicator's reserved bottom margin shrinks
the scene and lifts an absolutely-positioned in-scene button well above the
tab bar. The Home start-session FAB doesn't have this problem because it's a
direct absolute child of `ScreenWrapper`, anchored to the screen rather than
to flex-laid-out content.

Match the Home behavior: render the search FAB at the `SocialScreen` level
(sibling of the TabView/OfflineIndicator), reusing the Home FAB's
`floatingActionButtonContainer` + `floatingActionButton` styles and the same
`bottom: bottomTabBarHeight + 16` formula. Removed the in-scene button and its
now-dead `goToSearchScreenButton` style / `goToSearchButtonOffset` variable.

Behavior note: the FAB is now visible on both Friends tabs (was Friend
Requests only), matching the always-visible Home FAB and letting users start
a friend search from either tab.
