+++ src/styles/variables.ts
@@ -231,7 +231,6 @@ export default {
   bottomTabBarCounterSize: 20,
   calendarHeaderHeight: 50,
   sessionsCalendarArrowWidth: 44,
-  goToSearchButtonOffset: 20,
   qrCodeScreenSizePercentage: 0.7,
   qrCodeMinSizeLargeScreen: 300,
